Removed wavywall powerup from 4-player mode + removed test folder
